Ability to back up and restore components.
Note: Deletion is getting to be very complicated.  Will need to rethink.

Right now deletion is in two phases.  Remove, which Removes the majority
of components, so queries are no longer met.  And Delete,
which absolutely positively Deletes everything on the component.

Only because of Backup.  We need to retain Backup.  And because we can't
rely on Delete anymore.
The restore system has to manually Delete the Backup on Restoration.

So it has become a bit confusing.  And we have a bug where Splats seem
to duplicate, or come back out of the aether.  When they are removed.

Bizarre.


Signed-off-by: James <james.a.forbes@gmail.com>