Using Acceleration more across the board.
Patrol is not a good fit for Splat.  Patrol applies an even
acceleration every step.  As it is determined for you to reach
a specific target.

Splat needs acceleration/deceleration properties that affect the entity
acceleration over time.  So that splat can have a large/small initial speed
with a controlled instant/gradual slow down.

And splat needs a sister system for reforming.  Yes, it isn't very general.
And perhaps it is not a separate system at all.  But sometimes forcing
things to be general just makes for shitty code.




Signed-off-by: James <james.a.forbes@gmail.com>