Replaced Share with OwnerOffset, allows you to share, but gives you the ability to move shared values by a given amount for a given property.  More useful, if a little verbose.
In the future I might add Shared again, and it is just sugar for
OwnerOffset:0

Now have a concept of Ownership.  But only the owned knows about the
relationship.  This means it is the childs responsibility to clean itself
up when the Parent dies, instead of the parents.

Hopefully this will make for more interesting customization with provincial
reactions to the Parent's state.  Instead of global, standard reactions.

When you die, your Pickup is gone, because it is not a part of the Backup.
This is desired behaviour.  But in future we may want to Backup our
inventory/items, things we own etc.  Hopefully this backup behaviour
can be specified on the child instead of the parent having to have
a link to all its children.

Much easier to keep track of the state of one parent, than tracking
all the states of all children.


Signed-off-by: James <james.a.forbes@gmail.com>